+++ e2e/actions/regressions.go
@@ -18,6 +18,7 @@ import (
 
 	"github.com/sylabs/singularity/e2e/internal/e2e"
 	"github.com/sylabs/singularity/internal/pkg/cache"
+	"github.com/sylabs/singularity/internal/pkg/remote"
 	"github.com/sylabs/singularity/internal/pkg/test/tool/require"
 	"github.com/sylabs/singularity/internal/pkg/util/fs"
 )
@@ -651,7 +652,7 @@ func (c actionTests) invalidRemote(t *testing.T) {
 	)
 
 	// Add another endpoint
-	argv = []string{"add", "--no-login", testEndpoint, testEndpointURI}
+	argv = []string{"add", "--no-login", "--no-default", testEndpoint, testEndpointURI}
 	c.env.RunSingularity(
 		t,
 		e2e.AsSubtest("remote add"),
@@ -662,6 +663,17 @@ func (c actionTests) invalidRemote(t *testing.T) {
 	)
 	// Remove test remote when we are done here
 	defer func(t *testing.T) {
+		// Set default back to SylabsCloud
+		argv = []string{"use", remote.DefaultRemoteName}
+		c.env.RunSingularity(
+			t,
+			e2e.AsSubtest("remote use"),
+			e2e.WithProfile(e2e.UserProfile),
+			e2e.WithCommand("remote"),
+			e2e.WithArgs(argv...),
+			e2e.ExpectExit(0),
+		)
+
 		argv := []string{"remove", testEndpoint}
 		c.env.RunSingularity(
 			t,