+++ internal/app/singularity/keyserver_logout.go
@@ -29,10 +29,6 @@ func KeyserverLogout(usrConfigFile, name string) (err error) {
 		return fmt.Errorf("while parsing remote config data: %s", err)
 	}
 
-	if err := syncSysConfig(c); err != nil {
-		return err
-	}
-
 	// services
 	if err := c.Logout(name); err != nil {
 		return fmt.Errorf("while verifying token: %v", err)